Building Bonus tweaks + new way of getting metal
BONUSES:
Added food bonus of 5 to saddler to simulate more effective saddles and harnesses being used in farming. 
Added food bonus of 10 to Merchants Quarter. I found it a bit odd that merchants would stop trading food at a fairground.

PRODUCTION
Added metal production of 1 to fairground - This is a way for people far removed to gain access to metals, albeit in small quantities.
     - To balance this, it now produces 1 less of goods and wealth - Traders gotta trade.
Added metal production of 2 to Merchant's quarter
     - Produces 2 less of goods and 5 less of wealth